Rework grid logic, implement grid registery, service, extensions. Rework building place check logic
