+++ src/DiplomaGame/Assets/Config/Buildings/ConfigLoader.cs
@@ -1,5 +1,4 @@
-﻿using Biomes;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 using UnityEngine;
