+++ src/DiplomaGame/Assets/Config/InputSystem/ConfigLoader.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using UnityEngine;
 
 public static partial class ConfigLoader
