+++ src/DiplomaGame/Assets/Config/Map/MapConfig.json
@@ -1,6 +1,6 @@
 ﻿{
-	"MapWidth": 1000.0,
-	"MapHeight": 1000.0,
+	"MapWidth": 50.0,
+	"MapHeight": 50.0,
 	"MapStartPointX": 0.0,
 	"MapStartPointY": 0.0,
 	"CellSize": 10.0
