+++ src/DiplomaGame/Assets/Extensions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69232b161ba7f2440a8c444ed80b9a70
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
