+++ src/DiplomaGame/Assets/GameUtilities/Assets.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace GameUtilities
 {
