+++ src/DiplomaGame/Assets/Grid/GridRegistry.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f240a01af99c1934b85c80d014593e19
+guid: da8effc999384954d808fb56f78f27fd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
