+++ src/DiplomaGame/Assets/Grid/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1bfcdaf59e6e0d241ba90e85fbd6dca6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
