+++ src/DiplomaGame/Assets/Grid/Interfaces/IBaseGrid.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using UnityEngine;
+
+public interface IBaseGrid
+{
+    int Width { get; }
+    int Height { get; }
+    float CellSize { get; }
+    Vector3 OriginPosition { get; }
+    Type GridObjectType { get; }
+}
