+++ src/DiplomaGame/Assets/Grid/Interfaces/IBaseGrid.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfa214a680535e542aefd2a1caba5fea
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
