+++ src/DiplomaGame/Assets/Grid/Interfaces/IGridObject.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+public interface IGridObject
+{
+    int X { get; }
+    int Y { get; }
+    Guid Guid { get; }
+    Guid GetGuid();
+}
