+++ src/DiplomaGame/Assets/Grid/Interfaces/IGridObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ef301e9a0eeaa34a9278738fb3e4b21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
