+++ src/DiplomaGame/Assets/Grid/Interfaces/ITypedGrid.cs
@@ -1,0 +1,5 @@
+﻿public interface ITypedGrid : IBaseGrid
+{
+    IGridObject GetGridObjectInterface(int x, int y);
+}
+
