+++ src/DiplomaGame/Assets/Grid/Interfaces/ITypedGrid.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8a6d617603edab4899e90f52b821814
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
