+++ src/DiplomaGame/Assets/Grid/MapGridEventArgs/OnGridValueChangedEventArgs.cs
@@ -1,0 +1,6 @@
+using System;
+public class OnGridValueChangedEventArgs : EventArgs
+{
+	public int X { get; set; }
+    public int Y { get; set; }
+}
