+++ src/DiplomaGame/Assets/Grid/TypedGridAdapter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cdd70b5a11bf896479636d5f99fcbe9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
