--- src/DiplomaGame/Assets/Map/MapGridEventArgs/OnGridValueChangedEventArgs.cs
@@ -1,6 +1,0 @@
-using System;
-public class OnGridValueChangedEventArgs : EventArgs
-{
-	public int X { get; set; }
-    public int Y { get; set; }
-}
