+++ src/DiplomaGame/Assets/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f617a8f291133243b228d12d19891dc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
