+++ src/DiplomaGame/Assets/Services/GridService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f733412f3d055224ab3f5bfb4f6b5b05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
