+++ src/DiplomaGame/Assets/Tilemap/Tilemap.cs
@@ -2,10 +2,10 @@ using UnityEngine;
 
 public class Tilemap
 {
-	private MapGrid<TilemapObject> _grid;
+	private MapGrid<TilemapGridObject> _grid;
 
 	public Tilemap(int weight, int height, float cellSize, Vector3 originPosition) {
-		_grid = new MapGrid<TilemapObject>(weight, height, cellSize, originPosition, (g, x, y) => new TilemapObject(g, x, y, new TilemapSprite(TerrainTexturesConfig.TerrainTextures[0].Id)));
+		_grid = new MapGrid<TilemapGridObject>(weight, height, cellSize, originPosition, (g, x, y) => new TilemapGridObject(g, x, y, new TilemapSprite(TerrainTexturesConfig.TerrainTextures[0].Id)));
 	}
 
 	public void SetTilemapSprite(Vector3 wordPosition, TilemapSprite tilemapSprite)
