+++ src/DiplomaGame/Assets/Tilemap/TilemapGridObject.cs.meta
